Big refactoring: replaced AsyncStream with a simpler mechanism of returning the captured image
